Correções no código no uso do EJB. A persistencia da venda continua com erros ao tentar persistir o Item

+++ src/java/JPA/PerfilJpaController.java
@@ -6,18 +6,22 @@
 
 package JPA;
 
+import JPA.exceptions.IllegalOrphanException;
 import JPA.exceptions.NonexistentEntityException;
 import JPA.exceptions.RollbackFailureException;
 import java.io.Serializable;
+import javax.persistence.Query;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Root;
+import model.Funcionario;
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceUnit;
-import javax.persistence.Query;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import model.Perfil;
 
 /**
@@ -26,23 +30,38 @@ import model.Perfil;
  */
 @Stateless
 public class PerfilJpaController implements Serializable {
+
     @PersistenceUnit(unitName = "BazarWebPU") //inject from your application server
     private EntityManagerFactory emf;
-    
+
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
     public void create(Perfil perfil) throws RollbackFailureException, Exception {
+        if (perfil.getFuncionarioCollection() == null) {
+            perfil.setFuncionarioCollection(new ArrayList<Funcionario>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
-            em.persist(perfil);
-        } catch (Exception ex) {
-            try {
-            } catch (Exception re) {
-                throw new RollbackFailureException("An error occurred attempting to roll back the transaction.", re);
+            Collection<Funcionario> attachedFuncionarioCollection = new ArrayList<Funcionario>();
+            for (Funcionario funcionarioCollectionFuncionarioToAttach : perfil.getFuncionarioCollection()) {
+                funcionarioCollectionFuncionarioToAttach = em.getReference(funcionarioCollectionFuncionarioToAttach.getClass(), funcionarioCollectionFuncionarioToAttach.getId());
+                attachedFuncionarioCollection.add(funcionarioCollectionFuncionarioToAttach);
             }
+            perfil.setFuncionarioCollection(attachedFuncionarioCollection);
+            em.persist(perfil);
+            for (Funcionario funcionarioCollectionFuncionario : perfil.getFuncionarioCollection()) {
+                Perfil oldPerfilIdOfFuncionarioCollectionFuncionario = funcionarioCollectionFuncionario.getPerfilId();
+                funcionarioCollectionFuncionario.setPerfilId(perfil);
+                funcionarioCollectionFuncionario = em.merge(funcionarioCollectionFuncionario);
+                if (oldPerfilIdOfFuncionarioCollectionFuncionario != null) {
+                    oldPerfilIdOfFuncionarioCollectionFuncionario.getFuncionarioCollection().remove(funcionarioCollectionFuncionario);
+                    oldPerfilIdOfFuncionarioCollectionFuncionario = em.merge(oldPerfilIdOfFuncionarioCollectionFuncionario);
+                }
+            }
+        } catch (Exception ex) {
             throw ex;
         } finally {
             if (em != null) {
@@ -51,16 +70,45 @@ public class PerfilJpaController implements Serializable {
         }
     }
 
-    public void edit(Perfil perfil) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Perfil perfil) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
-            perfil = em.merge(perfil);
-        } catch (Exception ex) {
-            try {
-            } catch (Exception re) {
-                throw new RollbackFailureException("An error occurred attempting to roll back the transaction.", re);
+            Perfil persistentPerfil = em.find(Perfil.class, perfil.getId());
+            Collection<Funcionario> funcionarioCollectionOld = persistentPerfil.getFuncionarioCollection();
+            Collection<Funcionario> funcionarioCollectionNew = perfil.getFuncionarioCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Funcionario funcionarioCollectionOldFuncionario : funcionarioCollectionOld) {
+                if (!funcionarioCollectionNew.contains(funcionarioCollectionOldFuncionario)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Funcionario " + funcionarioCollectionOldFuncionario + " since its perfilId field is not nullable.");
+                }
             }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Funcionario> attachedFuncionarioCollectionNew = new ArrayList<Funcionario>();
+            for (Funcionario funcionarioCollectionNewFuncionarioToAttach : funcionarioCollectionNew) {
+                funcionarioCollectionNewFuncionarioToAttach = em.getReference(funcionarioCollectionNewFuncionarioToAttach.getClass(), funcionarioCollectionNewFuncionarioToAttach.getId());
+                attachedFuncionarioCollectionNew.add(funcionarioCollectionNewFuncionarioToAttach);
+            }
+            funcionarioCollectionNew = attachedFuncionarioCollectionNew;
+            perfil.setFuncionarioCollection(funcionarioCollectionNew);
+            perfil = em.merge(perfil);
+            for (Funcionario funcionarioCollectionNewFuncionario : funcionarioCollectionNew) {
+                if (!funcionarioCollectionOld.contains(funcionarioCollectionNewFuncionario)) {
+                    Perfil oldPerfilIdOfFuncionarioCollectionNewFuncionario = funcionarioCollectionNewFuncionario.getPerfilId();
+                    funcionarioCollectionNewFuncionario.setPerfilId(perfil);
+                    funcionarioCollectionNewFuncionario = em.merge(funcionarioCollectionNewFuncionario);
+                    if (oldPerfilIdOfFuncionarioCollectionNewFuncionario != null && !oldPerfilIdOfFuncionarioCollectionNewFuncionario.equals(perfil)) {
+                        oldPerfilIdOfFuncionarioCollectionNewFuncionario.getFuncionarioCollection().remove(funcionarioCollectionNewFuncionario);
+                        oldPerfilIdOfFuncionarioCollectionNewFuncionario = em.merge(oldPerfilIdOfFuncionarioCollectionNewFuncionario);
+                    }
+                }
+            }
+        } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
                 Integer id = perfil.getId();
@@ -76,7 +124,7 @@ public class PerfilJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -87,12 +135,19 @@ public class PerfilJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The perfil with id " + id + " no longer exists.", enfe);
             }
+            List<String> illegalOrphanMessages = null;
+            Collection<Funcionario> funcionarioCollectionOrphanCheck = perfil.getFuncionarioCollection();
+            for (Funcionario funcionarioCollectionOrphanCheckFuncionario : funcionarioCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Perfil (" + perfil + ") cannot be destroyed since the Funcionario " + funcionarioCollectionOrphanCheckFuncionario + " in its funcionarioCollection field has a non-nullable perfilId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
             em.remove(perfil);
         } catch (Exception ex) {
-            try {
-            } catch (Exception re) {
-                throw new RollbackFailureException("An error occurred attempting to roll back the transaction.", re);
-            }
             throw ex;
         } finally {
             if (em != null) {
